adds header image
