+++ App.js
@@ -5,12 +5,14 @@ import { createNativeStackNavigator } from '@react-navigation/native-stack';
 import LoginScreen from './navigation/components/LoginScreen';
 import RegistrationForm from './navigation/components/RegistrationForm';
 import MainContainer from './navigation/MainContainer';
+import Bibliotech from './navigation/components/Bibliotech';
 // import AddBookManually from './navigation/components/AddBookManually';
-import Fetch from './src/Fetch';
 const Stack = createNativeStackNavigator();
 
 export default function App() {
   return (
+    <View style={{ flex: 1}}>
+    <Bibliotech />
     <NavigationContainer>
       <Stack.Navigator>
         <Stack.Screen
@@ -27,12 +29,9 @@ export default function App() {
           options={{ headerShown: false }}
           name="RegistrationForm"
           component={RegistrationForm} />
-      {/* <Stack.Screen
-          options={{ headerShown: false }}
-          name="AddBookManually"
-          component={AddBookManually} /> */}
       </Stack.Navigator>
     </NavigationContainer>
+    </View>
   );
 };
