+++ navigation/components/ChatScreen.js
@@ -1,0 +1,20 @@
+import React from "react";
+import { View, Text, StyleSheet } from 'react-native';
+
+export default class ChatScreen extends React.Component {
+    render() {
+        return (
+            <View style={styles.container}>
+                <Text>Screen</Text>
+            </View>
+        );
+    }
+}
+
+const styles = StyleSheet.create({
+    container: {
+        flex: 1,
+        justifyContent: "center",
+        alignItems: "center"
+    }
+});
